\name{Adstock}
\alias{Adstock}
\title{Adstock}
\usage{
  AdStock(df,field,rate = 0.9)
}
\description{

Calcula el Adstock con un decay indicado

}

\arguments{
  \item{df}{Es el dataframe de la database}

  \item{field}{Es el número del campo de la variable objetivo}

  \item{rate}{Es el decay seleccionado}
}

\examples{
  AdStock(Base,35,0.7)
}


